Added option to completely bypass NAG libraries
In the case you don't have a NAG license, you can use the -1 flag for
the NAG argument in order to NOT use proprietary libraries. This
ofcourse has the caveat that there will no loger be a variance
reduction. You'll need to implement a variate reduction technique, e.g.
Antithetic Variates, instead of Quasi-random numbers.